modify class methods & fix errors
