+++ src/config.json
@@ -1,0 +1,6 @@
+{
+    "ORIG_DIR": "./etherscan/",
+    "COMPILED_DIR": "./compile_output/",
+    "COMPILED_DIR_OPT": "./compile_output_opt/",
+    "LOG_DIR": "./log/"
+}
