--- src/config.py
@@ -1,29 +1,0 @@
-# Configure all imports and project parameters
-
-import json
-import os
-import re
-from   pathlib import Path
-from   typing import Optional
-import solcx
-from   random import sample
-import gensim
-from   gensim.models.doc2vec import TaggedLineDocument
-from   gensim.models.doc2vec import TaggedDocument
-import torch
-import random
-import matplotlib.pyplot as plt
-import seaborn as sns
-from   sklearn.decomposition import PCA
-import pandas as pd
-from   tqdm import tqdm
-import subprocess
-import csv
-import adjustText
-from   typing import Dict
-
-Dict = {
-    "proj_path": "./SolcExample",
-    "bin_file":  "Payable.bin",
-    "save_path": "contract_out"
-}
